+++ android/src/main/java/com/reactnativestripesdk/StripeSdkPackage.kt
@@ -4,6 +4,7 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
+import com.reactnativestripesdk.addresssheet.AddressSheetViewManager
 import com.reactnativestripesdk.pushprovisioning.AddToWalletButtonManager
 
 class StripeSdkPackage : ReactPackage {
@@ -18,7 +19,8 @@ class StripeSdkPackage : ReactPackage {
         StripeContainerManager(),
         CardFormViewManager(),
         GooglePayButtonManager(),
-        AddToWalletButtonManager(reactContext)
+        AddToWalletButtonManager(reactContext),
+        AddressSheetViewManager()
       )
     }
 }