+++ example/src/screens/PaymentsUICompleteScreen.tsx
@@ -1,10 +1,13 @@
 import React, { useState } from 'react';
 import { Alert } from 'react-native';
 import {
+  AddressDetails,
   useStripe,
   BillingDetails,
   Address,
   PaymentSheetError,
+  AddressSheet,
+  AddressSheetError,
 } from '@stripe/stripe-react-native';
 import Button from '../components/Button';
 import PaymentScreen from '../components/PaymentScreen';
@@ -16,6 +19,7 @@ export default function PaymentsUICompleteScreen() {
     useStripe();
   const [paymentSheetEnabled, setPaymentSheetEnabled] = useState(false);
   const [loading, setLoading] = useState(false);
+  const [addressSheetVisible, setAddressSheetVisible] = useState(false);
   const [clientSecret, setClientSecret] = useState<string>();
 
   const fetchPaymentSheetParams = async () => {
@@ -58,7 +62,7 @@ export default function PaymentsUICompleteScreen() {
     setLoading(false);
   };
 
-  const initialisePaymentSheet = async () => {
+  const initialisePaymentSheet = async (shippingDetails?: AddressDetails) => {
     const { paymentIntent, ephemeralKey, customer } =
       await fetchPaymentSheetParams();
 
@@ -91,6 +95,7 @@ export default function PaymentsUICompleteScreen() {
       },
       returnURL: 'stripe-example://stripe-redirect',
       defaultBillingDetails: billingDetails,
+      defaultShippingDetails: shippingDetails,
       allowsDelayedPaymentMethods: true,
       appearance,
       primaryButtonLabel: 'purchase! 😃',
@@ -115,13 +120,58 @@ export default function PaymentsUICompleteScreen() {
     // To reduce loading time, make this request before the Checkout button is tapped, e.g. when the screen is loaded.
     <PaymentScreen onInit={initialisePaymentSheet}>
       <Button
+        variant="default"
+        loading={loading}
+        disabled={!paymentSheetEnabled}
+        title="Add shipping"
+        onPress={() => setAddressSheetVisible(true)}
+      />
+      <Button
         variant="primary"
         loading={loading}
         disabled={!paymentSheetEnabled}
         title="Checkout"
         onPress={openPaymentSheet}
       />
-
+      <AddressSheet
+        visible={addressSheetVisible}
+        onSubmit={async (result) => {
+          setPaymentSheetEnabled(false);
+          setAddressSheetVisible(false);
+          console.log(JSON.stringify(result, null, 2));
+          await initialisePaymentSheet(result);
+        }}
+        onError={(err) => {
+          if (err.code === AddressSheetError.Failed) {
+            Alert.alert('There was an error.', 'Check the logs for details.');
+            console.log(err?.localizedMessage);
+          }
+          setAddressSheetVisible(false);
+        }}
+        presentationStyle={'popover'}
+        animationStyle={'flip'}
+        appearance={{}}
+        defaultValues={{
+          name: 'Michael Scott',
+          phone: '111-222-3333',
+          isCheckboxSelected: true,
+          address: {
+            country: 'United States',
+            line1: 'Dunder Mifflin',
+            postalCode: '12345',
+            city: 'Scranton',
+          },
+        }}
+        additionalFields={{
+          phoneNumber: 'required',
+          checkboxLabel: 'Send me lots of emails',
+        }}
+        // allowedCountries={['US', 'CA']}
+        // autocompleteCountries={['CA']}
+        primaryButtonTitle={'use this address'}
+        sheetTitle={'🧙‍♀️ custom title'}
+        googlePlacesApiKey={'this-api-key-wont-work'}
+      />
       <Button
         title="Reset customer"
         onPress={async () => {